requirements.txt
# Model dependencies (install torch separately with the right CUDA wheel)
# pip install torch torchvision --extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.3.0
torchvision
diffusers>=0.32.0
transformers>=4.45.0
accelerate>=0.33.0
sentencepiece

# Frontend / backend dependencies
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bleach==6.2.0
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.133.0
ffmpy==1.0.0
filelock==3.24.3
fsspec==2026.2.0
gradio==6.6.0
gradio_client==2.1.0
gradio_image_annotation==0.5.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
kaggle==1.6.17
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pip==26.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
semantic-version==2.10.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
text-unidecode==1.3
tomlkit==0.13.3
tqdm==4.67.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
wheel==0.46.3
